Move jmp message encoding to own file
Disable old tests for now
